+++ relay-event-schema/src/protocol/ourlog.rs
@@ -1,42 +1,33 @@
 use relay_protocol::{
     Annotated, Empty, Error, FromValue, IntoValue, Object, SkipSerialization, Value,
 };
+use std::fmt::{self, Display};
+use std::str::FromStr;
 
 use serde::ser::SerializeMap;
+use serde::{Serialize, Serializer};
 
 use crate::processor::ProcessValue;
-use crate::protocol::{SpanId, TraceId};
+use crate::protocol::{SpanId, Timestamp, TraceId};
 
 #[derive(Clone, Debug, Default, PartialEq, Empty, FromValue, IntoValue, ProcessValue)]
 #[metastructure(process_func = "process_ourlog", value_type = "OurLog")]
 pub struct OurLog {
-    /// Time when the event occurred.
-    #[metastructure(required = true, trim = false)]
-    pub timestamp_nanos: Annotated<u64>,
-
-    /// Time when the event was observed.
-    #[metastructure(required = true, trim = false)]
-    pub observed_timestamp_nanos: Annotated<u64>,
+    /// Timestamp when the log was created.
+    #[metastructure(required = true)]
+    pub timestamp: Annotated<Timestamp>,
 
     /// The ID of the trace the log belongs to.
-    #[metastructure(required = false, trim = false)]
+    #[metastructure(required = true, trim = false)]
     pub trace_id: Annotated<TraceId>,
+
     /// The Span id.
-    ///
     #[metastructure(required = false, trim = false)]
     pub span_id: Annotated<SpanId>,
 
-    /// Trace flag bitfield.
-    #[metastructure(required = false)]
-    pub trace_flags: Annotated<u64>,
-
-    /// This is the original string representation of the severity as it is known at the source
-    #[metastructure(required = false, max_chars = 32, pii = "true", trim = false)]
-    pub severity_text: Annotated<String>,
-
-    /// Numerical representation of the severity level
-    #[metastructure(required = false)]
-    pub severity_number: Annotated<i64>,
+    /// The log level.
+    #[metastructure(required = true)]
+    pub level: Annotated<OurLogLevel>,
 
     /// Log body.
     #[metastructure(required = true, pii = "true", trim = false)]
@@ -51,15 +42,29 @@ pub struct OurLog {
     pub other: Object<Value>,
 }
 
+impl OurLog {
+    pub fn attribute(&self, key: &str) -> Option<Value> {
+        Some(match self.attributes.value()?.get(key) {
+            Some(value) => match value.value() {
+                Some(v) => match v {
+                    AttributeValue::StringValue(s) => Value::String(s.clone()),
+                    AttributeValue::IntValue(i) => Value::I64(*i),
+                    AttributeValue::DoubleValue(f) => Value::F64(*f),
+                    AttributeValue::BoolValue(b) => Value::Bool(*b),
+                    _ => return None,
+                },
+                None => return None,
+            },
+            None => return None,
+        })
+    }
+}
+
 #[derive(Debug, Clone, PartialEq, ProcessValue)]
 pub enum AttributeValue {
-    #[metastructure(field = "string_value", pii = "true")]
     StringValue(String),
-    #[metastructure(field = "int_value", pii = "true")]
     IntValue(i64),
-    #[metastructure(field = "double_value", pii = "true")]
     DoubleValue(f64),
-    #[metastructure(field = "bool_value", pii = "true")]
     BoolValue(bool),
     /// Any other unknown attribute value.
     ///
@@ -154,90 +159,319 @@ impl Empty for AttributeValue {
 impl FromValue for AttributeValue {
     fn from_value(value: Annotated<Value>) -> Annotated<Self> {
         match value {
-            Annotated(Some(Value::String(value)), meta) => {
-                Annotated(Some(AttributeValue::StringValue(value)), meta)
+            Annotated(Some(Value::Object(mut object)), meta) => {
+                let attribute_type = object
+                    .remove("type")
+                    .and_then(|v| v.value().cloned())
+                    .and_then(|v| match v {
+                        Value::String(s) => Some(s),
+                        _ => None,
+                    });
+
+                let value = object.remove("value");
+
+                match (attribute_type.as_deref(), value) {
+                    (Some("string"), Some(Annotated(Some(Value::String(string_value)), _))) => {
+                        Annotated(Some(AttributeValue::StringValue(string_value)), meta)
+                    }
+                    (Some("int"), Some(Annotated(Some(Value::String(string_value)), _))) => {
+                        let mut meta = meta;
+                        if let Ok(int_value) = string_value.parse::<i64>() {
+                            Annotated(Some(AttributeValue::IntValue(int_value)), meta)
+                        } else {
+                            meta.add_error(Error::invalid("integer could not be parsed."));
+                            meta.set_original_value(Some(Value::Object(object)));
+                            Annotated(None, meta)
+                        }
+                    }
+                    (Some("int"), Some(Annotated(Some(Value::I64(_)), _))) => {
+                        let mut meta = meta;
+                        meta.add_error(Error::expected(
+                            "64 bit integers have to be represented by a string in JSON",
+                        ));
+                        meta.set_original_value(Some(Value::Object(object)));
+                        Annotated(None, meta)
+                    }
+                    (Some("double"), Some(Annotated(Some(Value::F64(double_value)), _))) => {
+                        Annotated(Some(AttributeValue::DoubleValue(double_value)), meta)
+                    }
+                    (Some("bool"), Some(Annotated(Some(Value::Bool(bool_value)), _))) => {
+                        Annotated(Some(AttributeValue::BoolValue(bool_value)), meta)
+                    }
+                    (Some(_), Some(Annotated(Some(Value::String(unknown_value)), _))) => {
+                        Annotated(Some(AttributeValue::Unknown(unknown_value)), meta)
+                    }
+                    _ => {
+                        let mut meta = meta;
+                        meta.add_error(Error::expected(
+                            "a valid attribute value (string, int, double, bool)",
+                        ));
+                        meta.set_original_value(Some(Value::Object(object)));
+                        Annotated(None, meta)
+                    }
+                }
             }
-            Annotated(Some(Value::I64(value)), meta) => {
-                Annotated(Some(AttributeValue::IntValue(value)), meta)
-            }
-            Annotated(Some(Value::F64(value)), meta) => {
-                Annotated(Some(AttributeValue::DoubleValue(value)), meta)
-            }
-            Annotated(Some(Value::Bool(value)), meta) => {
-                Annotated(Some(AttributeValue::BoolValue(value)), meta)
-            }
+            Annotated(None, meta) => Annotated(None, meta),
             Annotated(Some(value), mut meta) => {
-                meta.add_error(Error::expected(
-                    "a valid attribute value (string, int, double, bool)",
-                ));
+                meta.add_error(Error::expected("an object"));
                 meta.set_original_value(Some(value));
                 Annotated(None, meta)
             }
-            Annotated(None, meta) => Annotated(None, meta),
         }
     }
 }
 
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub enum OurLogLevel {
+    Trace,
+    Debug,
+    Info,
+    Warn,
+    Error,
+    Fatal,
+    /// Unknown status, for forward compatibility.
+    Unknown(String),
+}
+
+impl OurLogLevel {
+    fn as_str(&self) -> &str {
+        match self {
+            OurLogLevel::Trace => "trace",
+            OurLogLevel::Debug => "debug",
+            OurLogLevel::Info => "info",
+            OurLogLevel::Warn => "warn",
+            OurLogLevel::Error => "error",
+            OurLogLevel::Fatal => "fatal",
+            OurLogLevel::Unknown(s) => s.as_str(),
+        }
+    }
+}
+
+impl Display for OurLogLevel {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.as_str())
+    }
+}
+impl FromStr for OurLogLevel {
+    type Err = ParseOurLogLevelError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Ok(match s {
+            "trace" => OurLogLevel::Trace,
+            "debug" => OurLogLevel::Debug,
+            "info" => OurLogLevel::Info,
+            "warn" => OurLogLevel::Warn,
+            "error" => OurLogLevel::Error,
+            "fatal" => OurLogLevel::Fatal,
+            other => OurLogLevel::Unknown(other.to_owned()),
+        })
+    }
+}
+
+impl FromValue for OurLogLevel {
+    fn from_value(value: Annotated<Value>) -> Annotated<Self> {
+        match value {
+            Annotated(Some(Value::String(value)), mut meta) => {
+                match OurLogLevel::from_str(&value) {
+                    Ok(value) => Annotated(Some(value), meta),
+                    Err(err) => {
+                        meta.add_error(Error::invalid(err));
+                        meta.set_original_value(Some(value));
+                        Annotated(None, meta)
+                    }
+                }
+            }
+            Annotated(None, meta) => Annotated(None, meta),
+            Annotated(Some(value), mut meta) => {
+                meta.add_error(Error::expected("a level"));
+                meta.set_original_value(Some(value));
+                Annotated(None, meta)
+            }
+        }
+    }
+}
+
+impl IntoValue for OurLogLevel {
+    fn into_value(self) -> Value {
+        Value::String(self.to_string())
+    }
+
+    fn serialize_payload<S>(&self, s: S, _behavior: SkipSerialization) -> Result<S::Ok, S::Error>
+    where
+        Self: Sized,
+        S: Serializer,
+    {
+        Serialize::serialize(self.as_str(), s)
+    }
+}
+
+impl ProcessValue for OurLogLevel {}
+
+impl Empty for OurLogLevel {
+    #[inline]
+    fn is_empty(&self) -> bool {
+        false
+    }
+}
+
+/// An error used when parsing `OurLogLevel`.
+#[derive(Debug)]
+pub struct ParseOurLogLevelError;
+
+impl fmt::Display for ParseOurLogLevelError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "invalid our log level")
+    }
+}
+
+impl std::error::Error for ParseOurLogLevelError {}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+    use relay_protocol::SerializableAnnotated;
 
     #[test]
     fn test_ourlog_serialization() {
         let json = r#"{
-  "timestamp_nanos": 1544712660300000000,
-  "observed_timestamp_nanos": 1544712660300000000,
-  "trace_id": "5b8efff798038103d269b633813fc60c",
-  "span_id": "eee19b7ec3c1b174",
-  "severity_text": "Information",
-  "severity_number": 10,
-  "body": "Example log record",
-  "attributes": {
-    "boolean.attribute": {
-      "bool_value": true
-    },
-    "double.attribute": {
-      "double_value": 637.704
-    },
-    "int.attribute": {
-      "int_value": 10
-    },
-    "string.attribute": {
-      "string_value": "some string"
+            "timestamp": 1544719860.0,
+            "trace_id": "5b8efff798038103d269b633813fc60c",
+            "span_id": "eee19b7ec3c1b174",
+            "level": "info",
+            "body": "Example log record",
+            "attributes": {
+                "boolean.attribute": {
+                    "value": true,
+                    "type": "bool"
+                },
+                "sentry.severity_text": {
+                    "value": "info",
+                    "type": "string"
+                },
+                "sentry.severity_number": {
+                    "value": "10",
+                    "type": "int"
+                },
+                "sentry.observed_timestamp_nanos": {
+                    "value": "1544712660300000000",
+                    "type": "int"
+                },
+                "sentry.trace_flags": {
+                    "value": "10",
+                    "type": "int"
+                }
+            }
+        }"#;
+
+        let data = Annotated::<OurLog>::from_json(json).unwrap();
+        insta::assert_debug_snapshot!(data, @r###"
+        OurLog {
+            timestamp: Timestamp(
+                2018-12-13T16:51:00Z,
+            ),
+            trace_id: TraceId(
+                "5b8efff798038103d269b633813fc60c",
+            ),
+            span_id: SpanId(
+                "eee19b7ec3c1b174",
+            ),
+            level: Info,
+            body: "Example log record",
+            attributes: {
+                "boolean.attribute": BoolValue(
+                    true,
+                ),
+                "sentry.observed_timestamp_nanos": IntValue(
+                    1544712660300000000,
+                ),
+                "sentry.severity_number": IntValue(
+                    10,
+                ),
+                "sentry.severity_text": StringValue(
+                    "info",
+                ),
+                "sentry.trace_flags": IntValue(
+                    10,
+                ),
+            },
+            other: {},
+        }
+        "###);
+
+        insta::assert_json_snapshot!(SerializableAnnotated(&data), @r###"
+        {
+          "timestamp": 1544719860.0,
+          "trace_id": "5b8efff798038103d269b633813fc60c",
+          "span_id": "eee19b7ec3c1b174",
+          "level": "info",
+          "body": "Example log record",
+          "attributes": {
+            "boolean.attribute": {
+              "bool_value": true
+            },
+            "sentry.observed_timestamp_nanos": {
+              "int_value": 1544712660300000000
+            },
+            "sentry.severity_number": {
+              "int_value": 10
+            },
+            "sentry.severity_text": {
+              "string_value": "info"
+            },
+            "sentry.trace_flags": {
+              "int_value": 10
+            }
+          }
+        }
+        "###);
     }
-  }
-}"#;
 
-        let mut attributes = Object::new();
-        attributes.insert(
-            "string.attribute".into(),
-            Annotated::new(AttributeValue::StringValue("some string".into())),
-        );
-        attributes.insert(
-            "boolean.attribute".into(),
-            Annotated::new(AttributeValue::BoolValue(true)),
-        );
-        attributes.insert(
-            "int.attribute".into(),
-            Annotated::new(AttributeValue::IntValue(10)),
-        );
-        attributes.insert(
-            "double.attribute".into(),
-            Annotated::new(AttributeValue::DoubleValue(637.704)),
-        );
+    #[test]
+    fn test_invalid_int_attribute() {
+        let json = r#"{
+            "timestamp": 1544719860.0,
+            "trace_id": "5b8efff798038103d269b633813fc60c",
+            "span_id": "eee19b7ec3c1b174",
+            "level": "info",
+            "body": "Example log record",
+            "attributes": {
+                "sentry.severity_number": {
+                    "value": 10,
+                    "type": "int"
+                }
+            }
+        }"#;
 
-        let log = Annotated::new(OurLog {
-            timestamp_nanos: Annotated::new(1544712660300000000),
-            observed_timestamp_nanos: Annotated::new(1544712660300000000),
-            severity_number: Annotated::new(10),
-            severity_text: Annotated::new("Information".to_string()),
-            trace_id: Annotated::new(TraceId("5b8efff798038103d269b633813fc60c".into())),
-            span_id: Annotated::new(SpanId("eee19b7ec3c1b174".into())),
-            body: Annotated::new("Example log record".to_string()),
-            attributes: Annotated::new(attributes),
-            ..Default::default()
-        });
+        let data = Annotated::<OurLog>::from_json(json).unwrap();
 
-        assert_eq!(json, log.to_json_pretty().unwrap());
+        insta::assert_json_snapshot!(SerializableAnnotated(&data), @r###"
+        {
+          "timestamp": 1544719860.0,
+          "trace_id": "5b8efff798038103d269b633813fc60c",
+          "span_id": "eee19b7ec3c1b174",
+          "level": "info",
+          "body": "Example log record",
+          "attributes": {
+            "sentry.severity_number": null
+          },
+          "_meta": {
+            "attributes": {
+              "sentry.severity_number": {
+                "": {
+                  "err": [
+                    [
+                      "invalid_data",
+                      {
+                        "reason": "expected 64 bit integers have to be represented by a string in JSON"
+                      }
+                    ]
+                  ],
+                  "val": {}
+                }
+              }
+            }
+          }
+        }
+        "###);
     }
 }